update merge type typing
