+++ ymir/web/src/pages/task/copy/index.js
@@ -10,7 +10,7 @@ import { randomNumber } from "@/utils/number"
 import Breadcrumbs from "@/components/common/breadcrumb"
 import commonStyles from "../common.less"
 import Desc from "@/components/form/desc"
-import DatasetName from "@/components/form/items/datasetName"
+import DatasetName from "@/components/form/items/DatasetName"
 import Dataset from '@/components/form/option/Dataset'
 
 function Copy() {
